file seprated function created to clean code
